Updating all search test to use page object structure
