+++ src/applications/search/tests/e2e/01-required.cypress.spec.js
@@ -5,155 +5,81 @@
  * @testrailinfo groupId 2925
  * @testrailinfo runName TA-2.0-e2e
  */
-
-class SearchComponent {
-  enableDropdownComponent = () => {
-    cy.route({
-      method: 'GET',
-      status: 200,
-      url: '/v0/feature_toggles*',
-      response: {
-        data: {
-          features: [
-            {
-              name: 'search_dropdown_component_enabled',
-              value: true,
-            },
-          ],
-        },
-      },
-    });
-  };
-
-  prepareDropdownSearch = term => {
-    cy.visit('/');
-    cy.get('button.sitewide-search-drop-down-panel-button').click();
-    cy.get('#search-header-dropdown-input-field').click();
-    cy.get('#search-header-dropdown-input-field')
-      .should('exist')
-      .should('not.be.disabled')
-      .type(term, { force: true });
-  };
-
-  mockFetchSuggestions = () => {
-    cy.route({
-      method: 'GET',
-      status: 200,
-      url: 'v0/search_typeahead?query=benefits',
-      response: [
-        'benefits response 1',
-        'benefits response 2',
-        'benefits response 3',
-        'benefits response 4',
-        'benefits response 5',
-      ],
-    });
-    cy.route({
-      method: 'GET',
-      status: 200,
-      url: 'v0/search_typeahead?query=health',
-      response: [
-        'health response 1',
-        'health response 2',
-        'health response 3',
-        'health response 4',
-        'health response 5',
-      ],
-    });
-  };
-
-  axeTestPage = () => {
-    cy.injectAxe();
-    cy.axeCheck();
-  };
-
-  loadComponent = term => {
-    this.enableDropdownComponent();
-    this.mockFetchSuggestions();
-    this.prepareDropdownSearch(term);
-  };
-
-  confirmDropDown = () => {
-    cy.get('#search-header-dropdown-listbox').should('be.visible');
-    cy.get('#search-header-dropdown-listbox')
-      .children()
-      .should('have.length', 5);
-  };
-}
+import SearchComponent from '../page-object/searchComponent';
 
 describe('Site-wide Search functionality with search dropdown component enabled', () => {
   beforeEach(function() {
     cy.server();
   });
 
+  const healthSearchTerm = 'health';
+  const benefitsSearchTerm = 'benefits';
   const searchComponent = new SearchComponent();
 
   it('passes axe requirements - - C12121', () => {
-    searchComponent.loadComponent('benefits');
+    searchComponent.loadComponent(benefitsSearchTerm);
     searchComponent.confirmDropDown();
     searchComponent.axeTestPage();
   });
 
   it('shows suggestions when user input is present and typeahead is enabled - - C12122', () => {
-    searchComponent.loadComponent('benefits');
+    searchComponent.loadComponent(benefitsSearchTerm);
     searchComponent.confirmDropDown();
   });
 
   it('Focusing the search button hides user input - - C12123', () => {
-    searchComponent.loadComponent('benefits');
+    searchComponent.loadComponent(benefitsSearchTerm);
     searchComponent.confirmDropDown();
-    cy.get('[data-e2e-id="search-header-dropdown-submit-button"]').focus();
-    cy.get('#search-header-dropdown-listbox').should('not.exist');
+    searchComponent.confirmSearchFocusHidesInput();
   });
 
   it('Focusing the input field repopulates suggestions - - C12124', () => {
-    searchComponent.loadComponent('health');
+    searchComponent.loadComponent(healthSearchTerm);
     searchComponent.confirmDropDown();
-    cy.get('[data-e2e-id="search-header-dropdown-submit-button"]').focus();
-    cy.get('#search-header-dropdown-listbox').should('not.exist');
-    cy.get('#search-header-dropdown-input-field').focus();
+    searchComponent.confirmSearchFocusHidesInput();
+    searchComponent.focusOnInputField();
     searchComponent.confirmDropDown();
   });
 
   it('Clicking search button initiates search for input - C12125', () => {
-    searchComponent.loadComponent('health');
-    cy.get('[data-e2e-id="search-header-dropdown-submit-button"]').click();
-    cy.url().should('contain', '/search/?query=health');
+    searchComponent.loadComponent(healthSearchTerm);
+    searchComponent.clickSubmitButton();
+    searchComponent.checkIfUrlContains(`/search/?query=${healthSearchTerm}`);
   });
 
   it('Pressing enter (focus on input field) initiates search for input - C12126', () => {
-    searchComponent.loadComponent('health');
-    cy.get('#search-header-dropdown-input-field').type('{enter}');
-    cy.url().should('contain', '/search/?query=health');
+    searchComponent.loadComponent(healthSearchTerm);
+    searchComponent.clickEnterInInputField();
+    searchComponent.checkIfUrlContains(`/search/?query=${healthSearchTerm}`);
   });
 
   it('Pressing enter (focus on search button) initiates search for input - C12127', () => {
-    searchComponent.loadComponent('health');
-    cy.get('[data-e2e-id="search-header-dropdown-submit-button"]').click();
-    cy.url().should('contain', '/search/?query=benefits');
+    searchComponent.loadComponent(benefitsSearchTerm);
+    searchComponent.clickSubmitButton();
+    searchComponent.checkIfUrlContains(`/search/?query=${benefitsSearchTerm}`);
   });
 
   it('Pressing space (focus on search button) initiates search for input - C12128', () => {
-    searchComponent.loadComponent('health');
-    cy.get('[data-e2e-id="search-header-dropdown-submit-button"]').click();
-    cy.url().should('contain', '/search/?query=health');
+    searchComponent.loadComponent(healthSearchTerm);
+    searchComponent.clickSubmitButton();
+    searchComponent.checkIfUrlContains(`/search/?query=${healthSearchTerm}`);
   });
 
   it('Clicking a dropdown option initiates a search using the suggestion - C12129', () => {
-    searchComponent.loadComponent('health');
+    searchComponent.loadComponent(healthSearchTerm);
     searchComponent.confirmDropDown();
-    cy.get('#search-header-dropdown-option-3').click();
-    cy.url().should('contain', '/search/?query=health%20response%204');
+    searchComponent.clickTypeAheadItem();
+    searchComponent.checkIfUrlContains(
+      `/search/?query=${healthSearchTerm}%20response%204`,
+    );
   });
 
   it('Can use the arrow keys to navigate suggestions, and press enter to search using them - C12130', () => {
-    searchComponent.loadComponent('benefits');
+    searchComponent.loadComponent(benefitsSearchTerm);
     searchComponent.confirmDropDown();
-    cy.get('#search-header-dropdown-input-field')
-      .type('{downarrow}')
-      .type('{downarrow}')
-      .type('{downarrow}')
-      .type('{enter}');
-    cy.url().should('contain', '/search/?query=benefits%20response%203');
+    searchComponent.navigateSearchSuggestions();
+    searchComponent.checkIfUrlContains(
+      `/search/?query=${benefitsSearchTerm}%20response%203`,
+    );
   });
 });
