+++ src/applications/search/tests/page-object/searchComponent.js
@@ -1,0 +1,131 @@
+/**
+ * This uses the page object to enhance the writing of Cypress E2E tests by modularizing the code snippets
+ */
+
+class SearchComponent {
+  /** Loads this component for testing by setting up the page and entering a search term */
+  loadComponent = term => {
+    this.enableDropdownComponent();
+    this.mockFetchSuggestions();
+    this.prepareDropdownSearch(term);
+  };
+
+  /** Enables the dropdown component on the page */
+  enableDropdownComponent = () => {
+    cy.route({
+      method: 'GET',
+      status: 200,
+      url: '/v0/feature_toggles*',
+      response: {
+        data: {
+          features: [
+            {
+              name: 'search_dropdown_component_enabled',
+              value: true,
+            },
+          ],
+        },
+      },
+    });
+  };
+
+  /** Visits the va.gov page, clicks the Search field, and enters the search term */
+  prepareDropdownSearch = term => {
+    cy.visit('/');
+    cy.get('button.sitewide-search-drop-down-panel-button').click();
+    cy.get('#search-header-dropdown-input-field').click();
+    cy.get('#search-header-dropdown-input-field')
+      .should('exist')
+      .should('not.be.disabled')
+      .type(term, { force: true });
+  };
+
+  /** Mocks the query typeahead suggestions */
+  mockFetchSuggestions = () => {
+    cy.route({
+      method: 'GET',
+      status: 200,
+      url: 'v0/search_typeahead?query=benefits',
+      response: [
+        'benefits response 1',
+        'benefits response 2',
+        'benefits response 3',
+        'benefits response 4',
+        'benefits response 5',
+      ],
+    });
+    cy.route({
+      method: 'GET',
+      status: 200,
+      url: 'v0/search_typeahead?query=health',
+      response: [
+        'health response 1',
+        'health response 2',
+        'health response 3',
+        'health response 4',
+        'health response 5',
+      ],
+    });
+  };
+
+  /** Injects and adds Axe check */
+  axeTestPage = () => {
+    cy.injectAxe();
+    cy.axeCheck();
+  };
+
+  /** Opens the dropdown and checks its length is 5 */
+  confirmDropDown = () => {
+    cy.get('#search-header-dropdown-listbox').should('be.visible');
+    cy.get('#search-header-dropdown-listbox')
+      .children()
+      .should('have.length', 5);
+  };
+
+  /** Gets the input field dropdown, moves down three items, and presses Enter */
+  navigateSearchSuggestions = () => {
+    cy.get('#search-header-dropdown-input-field')
+      .type('{downarrow}')
+      .type('{downarrow}')
+      .type('{downarrow}')
+      .type('{enter}');
+  };
+
+  /** Focuses on the Search button and checks that the listbox disappears */
+  confirmSearchFocusHidesInput = () => {
+    cy.get('[data-e2e-id="search-header-dropdown-submit-button"]').focus();
+    cy.get('#search-header-dropdown-listbox').should('not.exist');
+  };
+
+  /** Presses enter in the Search input field */
+  clickEnterInInputField = () => {
+    cy.get('#search-header-dropdown-input-field').type('{enter}');
+  };
+
+  /** Click the Search submit button */
+  clickSubmitButton = () => {
+    cy.get('[data-e2e-id="search-header-dropdown-submit-button"]').click();
+  };
+
+  /** Focuses on the Search input field */
+  focusOnInputField = () => {
+    cy.get('#search-header-dropdown-input-field').focus();
+  };
+
+  /** Focuses on the Search submit button */
+  focusOnSubmitButton = () => {
+    cy.get('[data-e2e-id="search-header-dropdown-submit-button"]').focus();
+  };
+
+  /** Checks the current URL for the queryString */
+  checkIfUrlContains = queryString => {
+    cy.url().should('contain', queryString);
+  };
+
+  /** Clicks on the third item in the Search dropdown list */
+  clickTypeAheadItem = () => {
+    cy.get(`#search-header-dropdown-option-3`).click();
+  };
+}
+
+export default SearchComponent;
